Ai-Resume-Analyzer-main (1)/Ai-Resume-Analyzer-main/Resume analyzer/Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pdfplumber
groq
python-multipart
requests
httpx
scikit-learn
